Refactor SessionTest core stuff to SessionTestBase
I have some pending feature work that creates a
new set of SessionTests that use this common base.
