+++ UnitTests/SessionTest.cs
@@ -2,222 +2,24 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using NUnit.Framework;
 using System.Threading;
+using NUnit.Framework;
+using QuickFix.Fields;
 using QuickFix.Logger;
 using QuickFix.Store;
 
 namespace UnitTests;
 
 [TestFixture]
-public class SessionTest
+public class SessionTest : SessionTestBase
 {
-    private SessionTestSupport.MockResponder _responder = new();
-
-    private QuickFix.SessionID _sessionId = new("unset", "unset", "unset");
-    private QuickFix.SessionSettings _settings = new();
-    private SessionTestSupport.MockApplication _application = new();
-    private QuickFix.Session? _session = null;
-    private QuickFix.Session? _session2 = null;
-    private QuickFix.SettingsDictionary _config = new();
-    private SeqNumType _seqNum = 1;
     private readonly Regex _msRegex = new(@"\.[\d]{1,3}$");
     private readonly Regex _microsecondRegex = new(@"\.[\d]{1,6}$");
 
     [SetUp]
     public void Setup()
     {
-        _responder = new SessionTestSupport.MockResponder();
-        _sessionId = new QuickFix.SessionID("FIX.4.2", "SENDER", "TARGET");
-        _application = new SessionTestSupport.MockApplication();
-        _settings = new QuickFix.SessionSettings();
-
-        _config = new QuickFix.SettingsDictionary();
-        _config.SetBool(QuickFix.SessionSettings.PERSIST_MESSAGES, false);
-        _config.SetString(QuickFix.SessionSettings.CONNECTION_TYPE, "initiator");
-        _config.SetString(QuickFix.SessionSettings.START_TIME, "00:00:00");
-        _config.SetString(QuickFix.SessionSettings.END_TIME, "00:00:00");
-        _settings.Set(_sessionId, _config);
-
-        var logFactory = new NullLogFactory(); // use QuickFix.ScreenLogFactory(settings) if you need to see output
-
-        // acceptor
-        _session = new QuickFix.Session(false, _application, new MemoryStoreFactory(), _sessionId,
-            new QuickFix.DataDictionaryProvider(),new QuickFix.SessionSchedule(_config), 0, logFactory, new QuickFix.DefaultMessageFactory(), "blah");
-        _session.SetResponder(_responder);
-        _session.CheckLatency = false;
-
-        // initiator
-        _session2 = new QuickFix.Session(true, _application, new MemoryStoreFactory(), new QuickFix.SessionID("FIX.4.2", "OTHER_SENDER", "OTHER_TARGET"),
-            new QuickFix.DataDictionaryProvider(), new QuickFix.SessionSchedule(_config), 0, logFactory, new QuickFix.DefaultMessageFactory(), "blah");
-        _session2.SetResponder(_responder);
-        _session2.CheckLatency = false;
-
-        _seqNum = 1;
-    }
-
-    public void Logon()
-    {
-        SendLogon(new QuickFix.FIX42.Logon());
-    }
-
-    public void Logon40()
-    {
-        SendLogon(new QuickFix.FIX40.Logon());
-    }
-
-    private void SendLogon(QuickFix.Message msg, SeqNumType n = 0)
-    {
-        msg.Header.SetField(new QuickFix.Fields.TargetCompID(_sessionId.SenderCompID));
-        msg.Header.SetField(new QuickFix.Fields.SenderCompID(_sessionId.TargetCompID));
-        msg.Header.SetField(new QuickFix.Fields.MsgSeqNum(n == 0 ? _seqNum++ : n));
-        msg.Header.SetField(new QuickFix.Fields.SendingTime(DateTime.UtcNow));
-        msg.SetField(new QuickFix.Fields.HeartBtInt(1));
-        _session!.Next(msg.ConstructString());
-    }
-
-    public bool SENT_SEQUENCE_RESET()
-    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.SEQUENCE_RESET) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.SEQUENCE_RESET].Count > 0;
-    }
-
-    public bool SENT_RESEND_REQUEST()
-    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.RESEND_REQUEST) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Count > 0;
-    }
-
-    public bool RESENT()
-    {
-        if (_responder.Dups.Count == 0)
-            return false;
-
-        _responder.Dups.Dequeue();
-        return true;
-    }
-
-    public bool SENT_REJECT()
-    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.REJECT) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.REJECT].Count>0;
-    }
-
-	    public bool SENT_HEART_BEAT()
-	    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.HEARTBEAT) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.HEARTBEAT].Count > 0;
-	    }
-
-    public bool SENT_BUSINESS_REJECT()
-    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.BUSINESS_MESSAGE_REJECT) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.BUSINESS_MESSAGE_REJECT].Count > 0;
-    }
-
-    public bool SENT_BUSINESS_REJECT(int reason)
-    {
-        if (!SENT_BUSINESS_REJECT())
-            return false;
-
-        QuickFix.Message msg = _responder.MsgLookup[QuickFix.Fields.MsgType.BUSINESS_MESSAGE_REJECT].First();
-
-        if (!msg.IsSetField(QuickFix.Fields.Tags.BusinessRejectReason))
-            return false;
-
-        QuickFix.Fields.BusinessRejectReason reasonField = new QuickFix.Fields.BusinessRejectReason();
-        msg.GetField(reasonField);
-        return reasonField.Value == reason;
-    }
-
-    public bool SENT_LOGOUT()
-    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.LOGOUT) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.LOGOUT].Count > 0;
-    }
-
-    public bool SENT_NOS()
-    {
-        return _responder.MsgLookup.ContainsKey(QuickFix.Fields.MsgType.NEWORDERSINGLE) &&
-            _responder.MsgLookup[QuickFix.Fields.MsgType.NEWORDERSINGLE].Count > 0;
-    }
-
-    public bool DISCONNECTED()
-    {
-        return _responder.Disconnected;
-    }
-
-    public bool SENT_REJECT(int reason, int refTag)
-    {
-        if (!SENT_REJECT())
-            return false;
-
-        QuickFix.Message msg = _responder.MsgLookup[QuickFix.Fields.MsgType.REJECT].First();
-
-        if (!msg.IsSetField(QuickFix.Fields.Tags.SessionRejectReason))
-            return false;
-
-        QuickFix.Fields.SessionRejectReason reasonField = new QuickFix.Fields.SessionRejectReason();
-        msg.GetField(reasonField);
-        if(reasonField.Value != reason)
-            return false;
-
-        if (!msg.IsSetField(QuickFix.Fields.Tags.RefTagID))
-            return false;
-
-        QuickFix.Fields.RefTagID refTagField = new QuickFix.Fields.RefTagID();
-        msg.GetField(refTagField);
-        return refTagField.Value == refTag;
-    }
-
-    public QuickFix.FIX42.NewOrderSingle CreateNOSMessage(SeqNumType n)
-    {
-        QuickFix.FIX42.NewOrderSingle order = new QuickFix.FIX42.NewOrderSingle(
-            new QuickFix.Fields.ClOrdID("1"),
-            new QuickFix.Fields.HandlInst(QuickFix.Fields.HandlInst.MANUAL_ORDER),
-            new QuickFix.Fields.Symbol("IBM"),
-            new QuickFix.Fields.Side(QuickFix.Fields.Side.BUY),
-            new QuickFix.Fields.TransactTime(),
-            new QuickFix.Fields.OrdType(QuickFix.Fields.OrdType.LIMIT));
-
-        order.Header.SetField(new QuickFix.Fields.TargetCompID(_sessionId.SenderCompID));
-        order.Header.SetField(new QuickFix.Fields.SenderCompID(_sessionId.TargetCompID));
-        order.Header.SetField(new QuickFix.Fields.SendingTime(DateTime.UtcNow));
-        order.Header.SetField(new QuickFix.Fields.MsgSeqNum(n));
-        return order;
-    }
-
-    public void SendNOSMessage()
-    {
-        _session!.Next(CreateNOSMessage(_seqNum++).ConstructString());
-    }
-
-    public void SendNOSMessage(SeqNumType n)
-    {
-        _session!.Next(CreateNOSMessage(n).ConstructString());
-    }
-
-    public void SendResendRequest(SeqNumType begin, SeqNumType end)
-    {
-        SendTheMessage(new QuickFix.FIX42.ResendRequest(
-            new QuickFix.Fields.BeginSeqNo(begin),
-            new QuickFix.Fields.EndSeqNo(end)));
-    }
-
-    public void SendResendRequest40(SeqNumType begin, SeqNumType end)
-    {
-        SendTheMessage(new QuickFix.FIX40.ResendRequest(
-            new QuickFix.Fields.BeginSeqNo(begin),
-            new QuickFix.Fields.EndSeqNo(end)));
-    }
-
-    private void SendTheMessage(QuickFix.Message msg)
-    {
-        msg.Header.SetField(new QuickFix.Fields.TargetCompID(_sessionId.SenderCompID));
-        msg.Header.SetField(new QuickFix.Fields.SenderCompID(_sessionId.TargetCompID));
-        msg.Header.SetField(new QuickFix.Fields.MsgSeqNum(_seqNum++));
-
-        _session!.Next(msg.ConstructString());
+        BaseSetup();
     }
 
     [Test]
@@ -239,7 +41,6 @@ public class SessionTest
         Logon();
         SendNOSMessage();
         Assert.That(SENT_REJECT(QuickFix.Fields.SessionRejectReason.VALUE_IS_INCORRECT,54));
-
     }
 
     [Test]
@@ -305,7 +106,7 @@ public class SessionTest
         order.Header.SetField(new QuickFix.Fields.TargetCompID(_sessionId.TargetCompID));
         order.Header.SetField(new QuickFix.Fields.SenderCompID(_sessionId.SenderCompID));
 
-        SeqNumType[] gapStarts = new[] { 1UL, 5UL, 11UL }; // 1st gap  from seq num 1 to 2 is just the Logon message
+        SeqNumType[] gapStarts = new[] { 1UL, 5UL, 11UL }; // 1st gap from seq num 1 to 2 is just the Logon message
         SeqNumType[] gapEnds = new[] { 2UL, 8UL, 15UL };
         int orderCount = 0;
 
@@ -861,32 +662,32 @@ public class SessionTest
 
         // <= FIX41
         _session.GenerateResendRequest(QuickFix.FixValues.BeginString.FIX41, 125);
-        Assert.That(_responder.GetCount(QuickFix.Fields.MsgType.RESEND_REQUEST), Is.EqualTo(1));
-        var rr = _responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Dequeue();
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.BeginSeqNo), Is.EqualTo(100));
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.EndSeqNo), Is.EqualTo(999999));
+        Assert.That(_responder.GetCount(MsgType.RESEND_REQUEST), Is.EqualTo(1));
+        var rr = _responder.MsgLookup[MsgType.RESEND_REQUEST].Dequeue();
+        Assert.That(rr.GetInt(Tags.BeginSeqNo), Is.EqualTo(100));
+        Assert.That(rr.GetInt(Tags.EndSeqNo), Is.EqualTo(999999));
 
         // >= FIX42
         _session.GenerateResendRequest(QuickFix.FixValues.BeginString.FIX42, 125);
-        Assert.That(_responder.GetCount(QuickFix.Fields.MsgType.RESEND_REQUEST), Is.EqualTo(1));
-        rr = _responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Dequeue();
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.BeginSeqNo), Is.EqualTo(100));
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.EndSeqNo), Is.EqualTo(0));
+        Assert.That(_responder.GetCount(MsgType.RESEND_REQUEST), Is.EqualTo(1));
+        rr = _responder.MsgLookup[MsgType.RESEND_REQUEST].Dequeue();
+        Assert.That(rr.GetInt(Tags.BeginSeqNo), Is.EqualTo(100));
+        Assert.That(rr.GetInt(Tags.EndSeqNo), Is.EqualTo(0));
 
         // Max resend is set: request is greater than max resend
         _session.MaxMessagesInResendRequest = 100;
         _session.GenerateResendRequest(QuickFix.FixValues.BeginString.FIX42, 225);
-        Assert.That(_responder.GetCount(QuickFix.Fields.MsgType.RESEND_REQUEST), Is.EqualTo(1));
-        rr = _responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Dequeue();
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.BeginSeqNo), Is.EqualTo(100));
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.EndSeqNo), Is.EqualTo(199));
+        Assert.That(_responder.GetCount(MsgType.RESEND_REQUEST), Is.EqualTo(1));
+        rr = _responder.MsgLookup[MsgType.RESEND_REQUEST].Dequeue();
+        Assert.That(rr.GetInt(Tags.BeginSeqNo), Is.EqualTo(100));
+        Assert.That(rr.GetInt(Tags.EndSeqNo), Is.EqualTo(199));
 
         // Max resend is set: request is lesser than max resend
         _session.GenerateResendRequest(QuickFix.FixValues.BeginString.FIX42, 175);
-        Assert.That(_responder.GetCount(QuickFix.Fields.MsgType.RESEND_REQUEST), Is.EqualTo(1));
-        rr = _responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Dequeue();
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.BeginSeqNo), Is.EqualTo(100));
-        Assert.That(rr.GetInt(QuickFix.Fields.Tags.EndSeqNo), Is.EqualTo(174));
+        Assert.That(_responder.GetCount(MsgType.RESEND_REQUEST), Is.EqualTo(1));
+        rr = _responder.MsgLookup[MsgType.RESEND_REQUEST].Dequeue();
+        Assert.That(rr.GetInt(Tags.BeginSeqNo), Is.EqualTo(100));
+        Assert.That(rr.GetInt(Tags.EndSeqNo), Is.EqualTo(174));
     }
 
     [Test]
@@ -899,10 +700,9 @@ public class SessionTest
         Assert.That(_session!.NextTargetMsgSeqNum, Is.EqualTo(3));
 
         SendNOSMessage(6);
-        Assert.That(_responder.GetCount(QuickFix.Fields.MsgType.RESEND_REQUEST), Is.EqualTo(1));
+        Assert.That(_responder.GetCount(MsgType.RESEND_REQUEST), Is.EqualTo(1));
 
-        var resendRequest =
-            (QuickFix.FIX42.ResendRequest)_responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Dequeue();
+        var resendRequest = (QuickFix.FIX42.ResendRequest)_responder.MsgLookup[MsgType.RESEND_REQUEST].Dequeue();
         Assert.That(resendRequest.BeginSeqNo.Value, Is.EqualTo(3));
         Assert.That(resendRequest.EndSeqNo.Value, Is.EqualTo(0)); // 0 is infinity, but actually we just want until 5
 
@@ -913,8 +713,8 @@ public class SessionTest
         foreach (SeqNumType i in new[] { 3, 4, 5, 7, 8 }) {
             var msg = CreateNOSMessage(i);
             if (i < 7) {
-                msg.Header.SetField(new QuickFix.Fields.PossDupFlag(true));
-                msg.Header.SetField(new QuickFix.Fields.OrigSendingTime(DateTime.MinValue)); // (value doesn't matter for test)
+                msg.Header.SetField(new PossDupFlag(true));
+                msg.Header.SetField(new OrigSendingTime(DateTime.MinValue)); // (value doesn't matter for test)
                 _session.Next(msg.ConstructString());
             }
             _session.Next(msg.ConstructString());
@@ -935,10 +735,10 @@ public class SessionTest
         Assert.That(_session!.NextTargetMsgSeqNum, Is.EqualTo(3));
 
         SendNOSMessage(6);
-        Assert.That(_responder.GetCount(QuickFix.Fields.MsgType.RESEND_REQUEST), Is.EqualTo(1));
+        Assert.That(_responder.GetCount(MsgType.RESEND_REQUEST), Is.EqualTo(1));
 
         var resendRequest =
-            (QuickFix.FIX42.ResendRequest)_responder.MsgLookup[QuickFix.Fields.MsgType.RESEND_REQUEST].Dequeue();
+            (QuickFix.FIX42.ResendRequest)_responder.MsgLookup[MsgType.RESEND_REQUEST].Dequeue();
         Assert.That(resendRequest.BeginSeqNo.Value, Is.EqualTo(3));
         Assert.That(resendRequest.EndSeqNo.Value, Is.EqualTo(5));
 
@@ -949,8 +749,8 @@ public class SessionTest
         foreach (SeqNumType i in new[] { 3, 4, 5, 7, 8 }) {
             var msg = CreateNOSMessage(i);
             if (i < 7) {
-                msg.Header.SetField(new QuickFix.Fields.PossDupFlag(true));
-                msg.Header.SetField(new QuickFix.Fields.OrigSendingTime(DateTime.MinValue)); // (value doesn't matter for test)
+                msg.Header.SetField(new PossDupFlag(true));
+                msg.Header.SetField(new OrigSendingTime(DateTime.MinValue)); // (value doesn't matter for test)
                 _session.Next(msg.ConstructString());
             }
             _session.Next(msg.ConstructString());
